% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgg_db.R
\docType{data}
\name{bgg_db}
\alias{bgg_db}
\title{bgg_db}
\format{Data frame.}
\source{
Board Game Geek
}
\usage{
bgg_db
}
\description{
A data frame with around 5,000 top board game information from Board Game Geek,
collected at Jun., 2018. It contains 20 columns, including numes of board games,
number of players, play time, ratings, designers, etc.
}
\keyword{datasets}
