% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boardgamegeek-package.R
\docType{package}
\name{boardgamegeek}
\alias{boardgamegeek}
\alias{boardgamegeek-package}
\title{boardgamegeek: a collection of board game information}
\description{
This package contains data from the world's biggest board game website "Board Game Geek",
and provides functions for you to search your favourite board games.
}
\author{
Junbin ZHANG
}
