% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommand.R
\name{recommand_by_num_of_player}
\alias{recommand_by_num_of_player}
\alias{recommand_by_time}
\alias{recommand_by_age}
\title{Recommand board games}
\usage{
recommand_by_num_of_player(min_p, max_p)

recommand_by_time(min_t, max_t)

recommand_by_age(min_a)
}
\arguments{
\item{min_p}{a number for minimum number of players.}

\item{max_p}{a number for maximum number of players.}

\item{min_t}{a number for minimum play time.}

\item{max_t}{a number for maximum play time.}

\item{min_a}{a number for minimum age.}
}
\value{
This function returns a data frame containing board games fulfill the searching requirements.
}
\description{
Recommand board games based on number of players, play time, or age range.
}
\examples{
recommand_by_num_of_player(1, 4)
recommand_by_time(0, 60)
recommand_by_age(12)

}
