% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_by_game_id}
\alias{search_by_game_id}
\alias{search_by_name}
\alias{search_by_year}
\alias{search_by_mechanic}
\alias{search_by_category}
\alias{search_by_designer}
\title{Search board games}
\usage{
search_by_game_id(id)

search_by_name(n)

search_by_year(y)

search_by_mechanic(m)

search_by_category(c)

search_by_designer(d)
}
\arguments{
\item{id}{a number for game ID.}

\item{n}{a string for name.}

\item{y}{a number for year.}

\item{m}{a string for mechanic.}

\item{c}{a string for category.}

\item{d}{a string for designer.}
}
\value{
This function returns a data frame containing board games fulfill the searching requirements.
}
\description{
Search board games by their ID, part of name, released year,
one of their mechanic, one of their category, or designer.
}
\examples{
search_by_game_id(1774430)
search_by_name("Gloomhaven")
search_by_year(2017)
search_by_mechanic("Role Playing")
search_by_category("Adventure")
search_by_designer("Isaac Childres")

}
