% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommand_by_age.R,
%   R/recommand_by_num_of_player.R, R/recommand_by_time.R
\name{recommand_by_age}
\alias{recommand_by_age}
\alias{recommand_by_num_of_player}
\alias{recommand_by_time}
\title{Recommand board games}
\usage{
recommand_by_age(min = NA)

recommand_by_num_of_player(min, max)
recommand_by_time(min, max)
recommand_by_age(age)

recommand_by_time(min = NA, max = NA)
}
\arguments{
\item{min}{a number.}

\item{max}{a number.}

\item{age}{a number to represent minimum age.}
}
\value{
This function returns a data frame containing board games fulfill the searching requirements.
}
\description{
Recommand board games based on number of players, play time, or age range.
}
\examples{
recommand_by_num_of_player(1, 4)
recommand_by_time(0, 60)
recommand_by_age(12)

}
