% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_by_category.R, R/search_by_designer.R,
%   R/search_by_game_id.R, R/search_by_mechanic.R, R/search_by_name.R,
%   R/search_by_year.R
\name{search_by_category}
\alias{search_by_category}
\alias{search_by_designer}
\alias{search_by_game_id}
\alias{search_by_mechanic}
\alias{search_by_name}
\alias{search_by_year}
\title{Search board games}
\usage{
search_by_category(c)

search_by_designer(d)

search_by_game_id(id)
search_by_name(name)
search_by_year(year)
search_by_mechanic(mechanic)
search_by_category(category)
search_by_designer(designer)

search_by_mechanic(m)

search_by_name(n)

search_by_year(y)
}
\arguments{
\item{id}{a number.}

\item{name}{a string.}

\item{year}{a number.}

\item{mechanic}{a string.}

\item{category}{a string.}

\item{designer}{a string.}
}
\value{
This function returns a data frame containing board games fulfill the searching requirements.
}
\description{
Search board games by their ID, part of name, released year,
one of their mechanic, one of their category, or designer.
}
\examples{
search_by_game_id(1774430)
search_by_name("Gloomhaven")
search_by_year(2017)
search_by_mechanic("Role Playing")
search_by_category("Adventure")
search_by_designer("Isaac Childres")

}
